Use OUT_OF_CODE_BLOCK for ResourceTypeClass field caching
This switches the cache to only invalidate when modifying
the Java structure (fields, class, method) or when
editing non-Java files like XML files. The out-of-code
mod count is what ClassInnerStuffCache uses.

For the most part, AS just uses the ClassInnerStuffCache
for R classes (which hit the augment). There are a few
random places where ResourceTypeClass is used:
AndroidInternalRClass, and R.java files. Even if
someone edits an R.java file, this seems like the right
mod count.

Bug: b/25078759
Change-Id: I77808773714ca96333437d07ef4ec985d4b8505a