IR: Inform user if pushing dex files fails
When pushing a dex file fails, we simply ignored it. This CL will
surface that error and terminate the launch. A subsequent launch will
then do a clean build.

A further optimization (not in this CL) would be to disable cold swap
for this device.

Fix for https://code.google.com/p/android/issues/detail?id=200752

Change-Id: Ia0823bd48d9af502ba659acbcd056c928cbddb7c