Perform full build if the app uses multiple processes
If the application uses the android:process attribute to launch
components in different processes, Instant Run will not work since we
wouldn't know to update the changes to multiple processes.

This CL simply turns off IR in such a case and does a full build.

This CL also cleans up the way the cause of a full build is reported.

Partial fix for
https://code.google.com/p/android/issues/detail?id=195907

Change-Id: I1ba23930c954e91dc2233e9557df3df904a88e2f