IR: Treat the app as dead when the executor is switched
When the executor is switched (from run to debug, say), the
LaunchTasksProviderFactory decides that it cannot do a hotswap since
there is no existing session. However, when setting the instant run
build options, we notice that the app is running and only create a
reload.dex.

This results in the deploy step missing out on the reload artifact (it
only knows about restart.dex, not reload.dex), and the user not seeing
the changes that were made.

Fix for https://code.google.com/p/android/issues/detail?id=200744

Change-Id: Ida927a18f5abc29efd66dad7f810d0642a7ec67a