[New PSD] Fixed transitive dependencies in 'Variants' view.
This CL fixes a bug where transitive dependencies are not shown
in the 'Variants' view. The reason is that the code is using
ArtifactDependencySpec, which includes the 'extension' and
'classifier' in its 'equals' and 'hashCode' implementations.

This is a problem because to find transitive dependencies we
need to match dependencies coming from build.gradle files, the
Gradle model and Maven POM files. All this sources may specify
the same dependency in a different way (e.g. build.gradle
file may not include a dependency's extension, while a POM
file does.)

The fix introduces PsdArtifactDependencySpec, which is similar
to the original ArtifactDependencySpec, but it ignores the
extension and classifier of a dependency.

Change-Id: Ie2102bec073d3493a3578a7eb929858ce1e8387a