Update AdapterProperty API for querying sync state
It turns out to be very useful to query an adapter property for
whether its in sync with its target property, as that allows us to
register validators against it, so we can, say, stop a user from
proceeding to the next step if their input is malformed.

By formalizing this behavior at the base class level, we can also
move some logic, previously handled by subclasses, to the base class.

Bug: http://b.android.com/201144
Change-Id: I3fedb72c06db543df2a1ba6250222f6de8c8a79e