If app is in bg, make sure app is terminated and re-launched
Currently, if an app is in the bg, we push the patch files via adb, and
then just relaunch the app, which results in the app showing up in the
foreground. However, if we had pushed cold swap dex files, they weren't
actually read because the app itself had not restarted - it was just
fronted.

This CL fixes the issue by making sure that the app is killed if it is
in the background.

Fixes https://code.google.com/p/android/issues/detail?id=201320

Change-Id: I26ae715fc33b89a33e1cb87b18c8e68201df6733
(cherry picked from commit ce86d48b920dac4dce26462d6690454e3a792b2e)