[New PSD] Fixed automatic selection sync in "Variants" view.
Problem:
Selecting a dependency in the "declared" dependencies view did
not select the "resolved" dependency in a reliable way. The issue
is that when the parent of a node to select was collapsed, the
node to select was completely ignored.

Fix:
This CL uses APIs in AbstractTreeBuilder to make node selection
reliable. Now if a node needs to be selected in the "Variants"
view, the parent node will be expanded, to display the newly
selected node.

Change-Id: Ica7082acc1735cbd5627a318f3ada3549030f39a