Allow SDK component install to be backgrounded
InstallSelectedPackagesStep gets a reference to the install wizard
dialog, and shows a "background" button. This cancells the wizard,
but allows the task install task to continue. Once the initial
phase has completed, it shows a notification if it has been
backgrounded.
The notification just kicks off the installation again, which will
start by finding and verifying the existing downloaded and unzipped
package, thus essentially skipping to the final steps.

Change-Id: Ifc08f7b9b564e7e7aab9f7bbbe5c64f50b6b4413