Switch from GLCanvas to GLJPanel.
GLCanvas is a heavy weight AWT component, which is causing issues on
OSX for Java versions 1.7 and later. This CL switches to using the
GLJPanel, which is the Swing alternative to GLCanvas. This CL also
contains a work around to the culling issues with GLJPanel, which was
the main reason GLCanvas was used in the first place.
Also, an extra glPointSize call, which made it's way in through a bad
merge, is removed.

Change-Id: Id27718b0c749b2bc24f4119ac1b68777c2f520a2