Use non Psi gradle build file merger when project is not init.
Implement GradleFileSimpleFileMerger based on GroovyLexer that
will not modify the Psi.
Use this simplified merger while the project is being created.
Continue to use the Psi merger when the project is initialized.

Add a test to verify that we can handle the merge with both
mergers. (Some tests have been slightly changed to get the same
results.)

Bug: https://b.corp.google.com/27514359

Change-Id: I48f2e52493b788a82630d570837420bdfafeb445