Fixed sync bug where using a newer builder-model throws exception.
This CL fixes a bug where syncing a project using a cached Gradle
model will fail in an installation of Android Studio using a new
version of the builder-model interfaces. Instead of sync failing
gracefully, an error message was displayed to the user.

The fix is to detect when a sync fails when using the cached
model, and requesting a full sync.

Change-Id: I13b03f06b25850de3342fecbdd5daeec4e875036