Support "Rerun" action
This CL adds a Rerun action to the toolbar and the Run menu.
In both places, it appears before the stop button. In the toolbar, it is
called "Rerun <config>", and in the main menu, it is called "Clean and
rerun <config>". It is only active when there is an existing run
session.

Despite its name, Rerun supports both run and debug sessions. Its
behavior is similar to the Instant Run/Debug actions: it doesn't show
the device picker and updates the app. However, it differs from them in
the sense that it terminates the app before pushing changes, so it
always does a cold swap.

When invoked from the menu, or if Shift key is pressed when the mouse is
clicked on the button, it performs a clean in addition to the terminate
and rerun actions.

Change-Id: I2d551caf9ca03acd9da8410d1f1a378f1c9309b2
(cherry picked from commit 5cdc08b484c3e5bd63708df4eac3007d52e86936)