Show debug sources from a different API level
This CL fixes a long standing issue: when debugging, we should show
sources that correspond to the API level of the device rather than the
compile or target SDK sources.

The sources corresponding to a location are obtained by
PositionManagerImpl.getPsiFileByLocation in a two step process:
first it finds the PsiClass corresponding to the given Location, and
then it identifies the navigation element corresponding to that class.
Currently, the PsiClass points to a class within the android.jar, and
its navigatable points to whatever sources are set for that jar.

This CL overrides this method. Whenever it detects that the source file
corresponds to the Android SDK, it then attempts to look at the current
device API level. It uses that information to retrieve the same file
present in the source folder corresponding to that API level.

Future enhancements:
 - We could allow switching between different sources (see
   AlternativeSourceProvider)

 - We could offer to download the corresponding sources

Fixes https://code.google.com/p/android/issues/detail?id=77315

Change-Id: I5e5390fb941aaf0b9839530070a8a9839c862464