Rename "ObservableProperty" to "AbstractProperty"
A renaming for this class has long been overdue. Originally it was
just called "Property", but that was considered to generic since the
term is so overloaded (e.g. Java System properties). So, to emphasize
that this property existed to be listened to, we called it
"ObservableProperty", especially is it inherited from
"ObservableValue".

However, later, a pattern emerged for naming read-only properties,
e.g.

BoolProperty -> ObservableBool
IntProperty -> ObservableInt
StringProperty -> ObservableString
OptionalProperty -> ObservableOptional
ObjectProperty -> ObservableObject

So, at this point, the base class name is confusing, because
"ObservableProperty" is very much NOT read-only. Therefore, I renamed
it in a way that hopefully conveys its a base class type while also
not simply being the generic "Property" name.

Change-Id: Ia6bdd50a19e55f8c8e6991fffc09e14f5313d368