Weak module reference in ModuleClassLoader
The static ModuleClassLoader cache holds on to the ModuleClassLoaders
which in turn holds on to the module. So, even if the keys (module) in
the cache are weakly held, they are never released, resulting in a
memory leak.

Hold only a weak reference to the module and check before accessing its
value.

Change-Id: I5acef6ca68cfbc50d7de83678e00aea45f3c51bf