Use separated inflate/rendering steps
This CL splits the RenderTask rendering in two different steps:
- Inflate: It should only be called when something is modified in the
  layout XML or resources.
- Render: Called to generate the bitmap from the inflated layout. This
  can also be now painted directly to a Graphics instance using
  RenderModel#paint

This allows to call render multiple times while only running the inflate
step once. This is useful if any property of the layout is modified
directly, like the layout params.
Now, ModelListener#modelChanged will be called when something is
modified in the XML or resources while ModelListener#modelRendered will
be called after a render has been requested and completed.

This CL also updates the frame time so we correctly preview the material
progress bar.

Change-Id: Ie5dd09e6da14d2dd2440509c1f3a21f760ce2189