Plumb some context through to CreateXmlResourcePanel
That way, some build systems can use the context to filter
down the number of base resource directory choices in
some cases (e.g., quickfix). In some cases there isn't
really any context (ChooseResourceDialog).

Also merge two of the constructors. Previously, one
of the constructors would always hide the value
and resource name fields. Add another flag to
control hiding the value field. There is also some
fallback behavior when the predefined name is null
where it would ignore the flag. Change one of the
callers to just pass in "true" for chooseName
when it is possible for the field to be null.

BUG: b/27775947
Change-Id: Ieede24fcf88744ec6bbe52d5304397c91bf49518