Correctly check whether project already has support annotations
After recent changes to artifact lookup, the code which checks
to see if you already depend on the support annotations would
check that you were using exactly the same version of the support
library or appcompat as the most recently available version (and
even required a dynamic dependency). That's not right; having
any version is okay, so we only check the group and artifact id's,
not the version.

Change-Id: I9a7e2949ca44c500150af7972c37a452cae7fe0f