Fix EditorFixture move caret methods
Following commit 067ca71435, the moveTo and moveToLine methods
were waiting for the breadcrumb panel in the editor to show before
clicking. That panel appears only for xml files, so instead we
wait for the event queue to be empty after opening a new file in
the editor, which waits for the breadcrumb panel to show if it
is present.

Change-Id: I19e3b2f4acc0ca28a8bddde12f7e252dc33baa35