[New PSD] Implemented "Speed Search" on Swing trees.
TreeSpeedSearch does not work with lazy-loading trees
backed by IDEA's TreeBuilder and TreeStructure. The
search only matches nodes that are 2 or 3 levels deep,
failing to show all the matching nodes.

This CL reimplements SpeedSearch for lazy-loading
trees and it correctly marks all the nodes that
match the given text pattern. It also expands parent
nodes to show search results if necessary.

Change-Id: Ia71b55608c8e0e573b3acd74e9ce619200ef8c75