Do waitForIdle after entering text through the robot
In GradleIncreaseLanguageLevelTest, tests fail because
editor.findOffset() is called before all the key events
triggered by editor.enterText() have been dealt with.
Add a waitForIdle at the end of Robot.enterText to ensure
that all the typing events have been handled properly before
continuing with the test.
For the moment, we create a new Robot that extends the FEST one
and overrides the enterText method. Once we make sure that the
issue is resolved, we will update the FEST robot.

Change-Id: Ibd3d73c801664b7eb4e6065d767fbfc530d83abb