Tests: EditorFixture.waitForCodeAnalysisHighlightCount fail
waitForCodeAnalysisHighlightCount can fail randomly if the total
number of errors is higher than the requested. For example, if the
method is called to check for 2 errors, but there are a total of 3,
90% of the time will start the analise with zero errors, loop, and
exit when reaching 2.
But sometimes, it is possible that the analise starts later, and
there are already 3, or that it's unlucky and the
loop misses the timing were there were 2.

waitForCodeAnalysisHighlightCount is now deprecated and its
behaviour is replaced by requireCodeAnalysisHighlightCount.
After all tests pass, there will be another CL to remove it.
All tests were run locally, but need to check on Jenkins.

Change-Id: I0d0ecbf965a94967959fc6713fd1337c870fea4e