Include lint specific error message in layout editor summary
The lint tooltip which shows the issue corresponding to a warning
icon in the design surface was only showing the static issue
description. That issue description is sometimes very generic
and you're not sure *what* in the layout actually triggered it.

This CL includes lint's error message too, before the explanation.
For an API error for example (where you're referencing an attribute
that requires a higher API level), this will include the name of
the attribute in question.

Change-Id: Id13a0e1c20989475da6a1a94ef144e27b857ccbf