Fixing scroll behavior with nested code scrollers.
As mousewheel events get consumed by the child scrollers in the code
samples, dispatch the event to the parent scroller in the following
conditions:
1) The code scroller is too small to need vertical scrollbars. Note
that this also fixes the default behavior of scrolling horizontally
in a vertical mouse wheel event IFF there is no vertical scrollbar.
2) The inner scrollbar has reached the extreem position in the
scroll direction.

This essentially makes the scroll panes act like html scrollers in
this context.

Change-Id: I4b5cc31508617f4e288ccde31ea1e1588c34933c