Fix for breadcrumb not working correctly after edit to atom stream.
The Breadcrumb used to keep a Path object for each segment, but this
meant it kept a capture ID, that could be old if a edit is made.

Now we keep a TreePath, this is always valid, as the TreePath contains
Nodes that are updated when a new capture ID is set.

Change-Id: I59ad2c41e7bd98a9199d698bc39151012480eb17