Pass modifier state to ViewActions
This allows view actions to change their behavior (and toolbar
appearance) based on the modifier keys. Note that right now
the modifier keys only affect the toolbar appearance if the mouse
is moving or if the design surface has the keyboard focus.

Also fix the selection handler such that you can use a modifier
key to toggle selection of an individual component instead of
replacing the whole selection.

Change-Id: If70a9dc9071fc9a5c3c30979acfd84e99cc38818