ProfilerPlugin, re-enables transform and supportlib if the flag is set.
The plugin is only applied if the experimental flag is set and so are
these features. If we move the plugin out of the experimental flag,
then at that point we would change the default of these features
depending on the flag. When the flag is set all the features
are on.

Change-Id: Ifb7441947bff459d5371da9e45a78ef4f83e3c06