Throw exception if creating a wizard without visible steps
Before, the wizard would try finish gracefully if created with
no visible steps, but in practice, this case shouldn't really happen.
After all, what does that mean if the user tries to open a dialog with
no steps? Should a window open and close immediately? Should a window
open with a blank body?

What's happening now is the dialog tries its best to open, shows the
first step in the wizard, and doesn't close. Hardly ideal...

It turns out the case of having no visible steps is almost certainly
an accident, so it's better to report it as an exception than try to
hide it silently.

Change-Id: Id929baa4d9bb35a4acf0e71bb1d2399cccd50d8d