Initial package refactoring in adt-ui
As the project is getting bigger, there is a need for organizing
the classes of adt-ui into packages. This CL starts this processes
and is based on internal discussions.

It's important to notice that even after this refactoring some classes
remain in the root of adt-ui. Therefore, there's still room for
improvement and follow up refactors might be done soon.

Change-Id: Ib455ad8fff69db3fd099d08c044b953f122b657d