209778: Rendering exception while creating layout-xlarge for settings activity
When creating resource overrides, make sure to use the folder type
of the file being forked, don't hardcode it for layout. This makes
it work from preference files and menu files, for example.

Also remove the hardcoded "Create Landscape" and "Create xlarge"
shortcut actions unless we're in a layout file.

Change-Id: I70dc99c3b6a6575b7c5a01cdcd9c1bfcb241d384