If user declines adding constraint layout, don't proceed with conversion
This CL changes the meaning of the return value of the
ensureLibraryIsIncluded method such that "false" no longer
means "didn't have import", it now means "didn't have import
and user declined adding it".  No code was actually looking at
the return value.

Use the return value in the conversion action such that if the
user attempts to convert a layout, and then declines the "Add
ConstraintLayout library to the project?" dialog, conversion
stops.

Change-Id: Ie9e57418095dca317146530bb4d350031192759e