[New PSD] Initial support for Java modules.
This CL adds initial support for showing
dependencies of Java modules.

It is disabled until at least it is possible
to add dependencies (to have feature parity
to the existing Android module support.)

Most of this CL handles refactoring of code
that was specific to Android modules and now
is generic enough that can be shared with
Java module support.

Change-Id: I3e3d235f50b5637046e44f1bc99a7b09d8429e6e
(cherry picked from commit 1dde09739a2cf7048b8725125388a3eeb9a4ad87)