Modified AxisComponent to support clamping to the next major tick
Added logic in the AxisComponent to update the Range's max target
during the animate/updataData phase to clamp to major ticks and to
sync to the parent axis. This requires moving the tick position/
scale calculations to the postAnimate phase. Because the AxisCompnent
can now take part in updating the Range's target, reordered the axes
in the Choreographer so that the Range objects interpolate after both
the LineChart and AxisComponents have their chance to update the
ranges' values.

Change-Id: I41781a7a9f3b30227ea7036b6c4ef00bf6f6cbf0