Fix GradleSyncTest.testModuleSelectionOnImport
There was a problem with the way the IDE requested focus: if the
IDE frame did not have a focused component, it would request the focus,
despite the fact that some other window (like a dialog) could have it.
Now, it checks that no-one has the focus before requesting it.

Change-Id: I52b3d33927562a1077be573a1a269a1b46fbab85