Fix IllegalArgException arising possibly due to duplicate keys
ImmutableMap.Builder throws an IllegalArgumentException when
constructing the map if it detects that there are duplicate keys. This
Cl avoids the issue by constructing an ImmutableMap as a copy of a
regular HashMap as opposed to using the Builder.

It is not entirely clear why there should be duplicate source folders
per API level though.

Fixes https://code.google.com/p/android/issues/detail?id=211788

Change-Id: Icb065846814fedf527e7d4a620a4c0695764157a