Fix NPE in DevicePicker (update invoked post dispose)
A recent change to DevicePicker nulled out the device list when the
dialog is disposed off. A pending runnable (for updating AVD list
for instance), could run into an NPE if it accessed the list.
This CL prevents the above scenario by guarding against a null list, and
making sure that the runnable proceeds only when it detects that the
dialog hasn't been disposed off yet.

Fixes https://code.google.com/p/android/issues/detail?id=211914

Change-Id: I054195776dde7c2e8042c9ad909b548b4e457c59