Re-introduce clean when timestamps don't match
An earlier commit (e296cedd8fc3c) removed unnecessary cleans and changed
the IDE to invoke a clean only when explicitly requested. It turns out
that some scenarios still need a clean. So this CL re-introduces clean
again. The behavior should match what was in 2.1. It may be cleaning in
more circumstances than needed, but that will be fixed in a subsequent
CL.

Fixes https://code.google.com/p/android/issues/detail?id=212041

Change-Id: Ia54a285bf446ebce7db881a4d69380e1de070e75