Extend the TestInvokeStrategy for use in general testing of Properties
The BatchInvoker falls back to using a Swing based asynchronous message
queue when used in the testing environment. This introduces the
potential for concurrency issues as callbacks are executed on a
different thread to the unit tests. There is also limited access to
the message queue to query whether all messages have been processed
or to inject assertions at points during processing. This change extends
TestInvokeStrategy to give tests control over asynchronous message
processing.

Change-Id: I44be28e7a68262919239fccb227a85e11dabf6e7