Zoom on design surface is now centered...
... Previously, when the ScreenView were zoomed the ViewPort position
were reset to 0,0 causing a weird zooming effect. The Zoom is now
centered on the  center of the viewport.
Also added a the possibility to zoom from 16% to 300% with the mouse wheel on the Zoom
and Pan widget. Limited the zoom to 300% otherwise it was too easy to
"overzoomed" and cause an out of memory error.

Change-Id: I8143e2e003afe463417411e05036fe3498f3c6dd