Change how test completion status is displayed.
When tests run to completion, the launch session is terminated with a
message "Tests ran to completion". Since this was passed as a message to
terminateLaunch, it appeared as red in the console, causing some
confusion.

This CL prints the message to stdout rather than stderr.

Fixes https://code.google.com/p/android/issues/detail?id=212331

Change-Id: Ie68c8051fb8b2015e732f016528a7acdee7e4ca6