Use a different configuration for a new ThemeResolver
When creating a new ThemeResolver we were setting the theme to null in
the configuration. This would cause that configuration to forget the
current theme and select the default one (usually the manifest theme).
This CL creates a new configuration to use in the ThemeResolver so the
main configuration is not affected.

Bug: http://b.android.com/213308
Change-Id: Ia41a19851576c4b4a46a60f2467d545bbd924dc3