Fix CPU usage label displaying Long.MAX_VALUE
getFormattedString() method of SingleUnitAxisFormatter was trying to
divide value by getMultiplier() (which returns 0), instead of
MULTIPLIER (which is 1). As a result, we were trying to divide some
value in the interval [0.0, 100.0] by 0 (which actually is a double very
close to 0). As a result, calling Math.round(...) on this division
returned Long.MAX_VALUE.

Change-Id: I685c334f6939cc31a08f56b14e12b522d87d7d2d