Add initial image diff test to adt-ui
The current UI tests in adt-ui package are very unreliable. They are
mostly visual and requires the developer to run a "proof of concept"
sample app that renders the components altogether and use their judgement
to check if everything is OK. This is, of course, not ideal, as a lot of
details might be missed and bugs could be inserted.

That being sad, it might be worth it to automatize this test process and
this CL starts doing that by adding unit tests to cover the components
of adt-ui, such as charts and axis.

Essentially, the tests builds a determined component, composed by
several Swing elements, and export it to an image. This image, in turn,
is compared against a target image. If the image differs more than a
certain threshold, the test should fail and the developer can verify the
inconsistency (hopefully) early enough to fix it. For now, the image
comparison is done by comparing each ARGB component of the images. Other
techniques (such as Perceptual Image Diff) might be used later.

The test classes (such as LineChartImageDiffTest) should reuse methods
and constants from ImageDiffUtil, which contains some common logic, such as
comparing images and exporting components to an image.

This CL is just an initial step and adds only the base class and a very
initial line chart tests, which contains only the test corresponding to
simple stacked line charts.

The idea is to cover all the line chart variation in this class and
create other test classes for other charts and components later. After
new components are added, it's also strongly recommended to add
corresponding tests for them in this package.

Change-Id: Ia02df515f9d8428f5328196f5d9ee11fef9d2bc0