Temporary fix of button lookup based on text (workaround on Mac.)
CL
https://googleplex-android-review.git.corp.google.com/1215572
broke button lookup based on text. This is broken only on Mac
laptops (other type of Mac seem to be OK.)

The problem is that on Mac laptops the text of buttons have
extra spaces, and the CL mentioned above removed a call to
'trim'. This CL reverts to the previous behavior, to
enable developers on Mac laptops run the UI tests locally.

Change-Id: Iefdd33dfb58f84f5d260786ad829d4e767170ffb