Gradle sync: cache exceptions during reproxying
There doesn't seem to be a strong need for keeping all instances of
org.gradle.tooling.model.UnsupportedMethodException, especially
when most of them have the same message and only differ in a few
reproxy() calls in the captured stacktrace.

The cache uses soft references because we don't want it to grow
arbitrarily large.

Change-Id: Ib72e65acdc9a0fc223e56f4acb8cfac5dae03ee3