Fix darcula colors in the layout editor
Using UiUtil.getListBackground() and UiUtil.getListForeground()
give different results whether AS is started in Light or Dark mode,
which means the colors are not right when switching from Light to
Dark mode.
Hardcode the dark version of the colors instead, to ensure they
have the expected value whatever mode AS is starting in.

Change-Id: I00210eb2a4addc8af95eccd86fdf749de1334ef9