Parser model: ability to add dependencies with excludes
This should allow us to manage androidTestCompile dependencies better.
For now there is not way to edit the excludes after they have been
added, the parser ignores the configuration block once it has been
written. Managing the block will be a big challenge and it's not clear
to me it's really worth it.

Bug: http://b.android.com/218935
Change-Id: Idec0f4baadc644f4132cc3dabedf3b69b7b01b83